+++ src/03-organisms/teasers-grid/stories/teasers-grid.data.js
@@ -1,182 +1,59 @@
+import {imageData } from '../../../01-atoms/image/stories/image.data'
+
+const img = {
+  dataSrcset: imageData.dataSrcset,
+  dataSrc: imageData.dataSrc,
+  src: imageData.src,
+  sizes: "(min-width: 720px) 30vw, 100vw"
+};
+
 export const teaserData = [
   {
-    img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/716x520",
-      src: "https://placehold.it/716x520",
-      sizes: "(min-width: 720px) 30vw, 100vw"
-    },
+    img,
     title: "Küchenabdeckungen"
   },
   {
-    img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/716x520",
-      src: "https://placehold.it/716x520",
-      sizes: "(min-width: 720px) 30vw, 100vw"
-    },
+    img,
     title: "Bodenbeläge und Treppen (Innen)"
   },
   {
-    img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/716x520",
-      src: "https://placehold.it/716x520",
-      sizes: "(min-width: 720px) 30vw, 100vw"
-    },
+    img,
     title: "Bodenbeläge und Treppen (Aussen)"
   },
   {
-    img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/716x520",
-      src: "https://placehold.it/716x520",
-      sizes: "(min-width: 720px) 30vw, 100vw"
-    },
+    img,
     title: "Elementbau (Innen)"
   },
   {
-    img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/716x520",
-      src: "https://placehold.it/716x520",
-      sizes: "(min-width: 720px) 30vw, 100vw"
-    },
+    img,
     title: "Elementbau (Aussen)"
   },
   {
-    img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/716x520",
-      src: "https://placehold.it/716x520",
-      sizes: "(min-width: 720px) 30vw, 100vw"
-    },
+    img,
     title: "Fassaden"
   },
   {
-    img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/716x520",
-      src: "https://placehold.it/716x520",
-      sizes: "(min-width: 720px) 30vw, 100vw"
-    },
+    img,
     title: "Gärten"
   },
   {
-    img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/716x520",
-      src: "https://placehold.it/716x520",
-      sizes: "(min-width: 720px) 30vw, 100vw"
-    },
+    img,
     title: "Grabsteine"
   },
   {
-    img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/716x520",
-      src: "https://placehold.it/716x520",
-      sizes: "(min-width: 720px) 30vw, 100vw"
-    },
+    img,
     title: "Berner Sandstein"
   },
   {
-    img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/716x520",
-      src: "https://placehold.it/716x520",
-      sizes: "(min-width: 720px) 30vw, 100vw"
-    },
+    img,
     title: "Oberflächenbearbeitung"
   },
   {
-    img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/716x520",
-      src: "https://placehold.it/716x520",
-      sizes: "(min-width: 720px) 30vw, 100vw"
-    },
+    img,
     title: "Badezimmer"
   },
   {
-    img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/716x520",
-      src: "https://placehold.it/716x520",
-      sizes: "(min-width: 720px) 30vw, 100vw"
-    },
+    img,
     title: "Wasseraufbereitung & Abwassertechnik"
   }
 ]