+++ src/04-templates/detail-template/index.js
@@ -8,7 +8,7 @@ import GoodToKnow from "../../03-organisms/good-to-know"
 import Teaser from "../../02-molecules/teaser"
 import { teaserData } from "../../03-organisms/teasers-grid/stories/teasers-grid.data"
 
-const imgData = Object.assign({}, teaserData[1].img, {src: 'https://placehold.it/1074x780'})
+const imgData = Object.assign({}, teaserData[1].img, {src: 'https://placehold.it/800x450'})
 
 const DetailTemplate = () => (
   <Layout classes={["t-template--overview"]}>